requirements_new.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
djangorestframework==3.16.0
django-htmx==1.19.0
channels==4.1.0
channels-redis==4.2.0
redis==5.2.1
idna==3.10
numpy==2.2.6
opencv-contrib-python==4.11.0.86
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
django-jazzmin==3.0.0
